Move definition of `Resource` type to model
